+++ src/renderer/screens/BoardsScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function BoardsScreen(): React.JSX.Element {
+  return <div>Boards & Keywords screen</div>;
+}
+
+export default BoardsScreen;