+++ src/renderer/screens/PipelineScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function PipelineScreen(): React.JSX.Element {
+  return <div>Pipeline screen</div>;
+}
+
+export default PipelineScreen;