+++ src/renderer/screens/ResultsScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ResultsScreen(): React.JSX.Element {
+  return <div>Results screen</div>;
+}
+
+export default ResultsScreen;