+++ src/renderer/screens/ResumeScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ResumeScreen(): React.JSX.Element {
+  return <div>Resume screen</div>;
+}
+
+export default ResumeScreen;