+++ src/renderer/screens/ScoutScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ScoutScreen(): React.JSX.Element {
+  return <div>Scout screen</div>;
+}
+
+export default ScoutScreen;