+++ src/renderer/screens/SettingsScreen.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function SettingsScreen(): React.JSX.Element {
+  return <div>Settings screen</div>;
+}
+
+export default SettingsScreen;